+++ marginaleffects/predictions.py
@@ -180,6 +180,8 @@ def predictions(
         out = get_z_p_ci(
             out, model, conf_level=conf_level, hypothesis_null=hypothesis_null
         )
+    else:
+        J = None
     out = get_transform(out, transform=transform)
     out = get_equivalence(out, equivalence=equivalence)
     out = sort_columns(out, by=by, newdata=newdata)
@@ -188,7 +190,7 @@ def predictions(
     if "rowid" in out.columns and pad.shape[0] > 0:
         out = out[: -pad.shape[0] :]
 
-    out = MarginaleffectsDataFrame(out, by=by, conf_level=conf_level, newdata=newdata)
+    out = MarginaleffectsDataFrame(out, by=by, conf_level=conf_level, jacobian=J, newdata=newdata)
     return out